Play_Sound
This is the function I built to play a random sound. The function works
but it freezes the game when it goes through the first loop and I am not
sure why.

I am also pushing all the music up to Git, I found an uncorrupted
version of bouncy1.

I also changed around the import for the sounds/music and made it
simplier.

I'm going to start working on the platform music now.